move deductions under nn
