--- src/aioway/__init__.py
@@ -1,22 +1,0 @@
-# Copyright (c) AIoWay Authors - All Rights Reserved
-
-"The global modules."
-
-from . import (
-    _api,
-    _costs,
-    _sess,
-    _thunks,
-    _utils,
-    api,
-    app,
-    deductions,
-    dsets,
-    emits,
-    errors,
-    modes,
-    rewrites,
-    schemas,
-    trainers,
-    tspecs,
-)
