+++ src/aioway/nn/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) AIoWay Authors - All Rights Reserved
+
+from .deductions import *
