+++ src/aioway/nn/deductions/deductions.py
