+++ src/aioway/nn/deductions/dense.py
