+++ src/aioway/nn/deductions/miscs.py
