+++ src/aioway/nn/deductions/sliding.py
